moved test into token-client and added test case with query parameters
